+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## Release 1.0.0-M20
 
 * #249 Remove APIs deprecated since 0.16
+* #250 Remove APIs deprecated since 0.17 
 
 ## Release 1.0.0-M19